gammatone proc is working fine

+++ codels/genom3_dataFiles.hpp
@@ -12,7 +12,7 @@
 
 using namespace openAFE;
 
-using twoCTypeBlockPtr = typename twoCTypeBlock<float>::twoCTypeBlockPtr;
+using twoCTypeBlockPtr = typename twoCTypeBlock<double>::twoCTypeBlockPtr;
 
 struct rosAFE_inputProcessors {
 	ProcessorVector< InputProc > processorsAccessor;
