+++ codels/src/Processors/inputProc.hpp
@@ -19,21 +19,21 @@
  
 namespace openAFE {
 
-	class InputProc : public TDSProcessor<float> {
+	class InputProc : public TDSProcessor<double> {
 					
 		private:
 			
 			bool in_doNormalize;
 			uint64_t in_normalizeValue;
 									
-			void process ( float* firstValue, size_t dim ) {
+			void process ( double* firstValue, size_t dim ) {
 				for ( unsigned int i = 0 ; i < dim ; ++i )
 					*( firstValue + i ) = *( firstValue + i ) / this->in_normalizeValue;
 			}
 
 		public:
 
-			InputProc ( const std::string nameArg, const uint32_t fs, const uint32_t bufferSize_s, bool in_doNormalize = true, uint64_t in_normalizeValue = MAXCODABLEVALUE ) : TDSProcessor<float> (nameArg, fs, fs, bufferSize_s, _inputProc) {
+			InputProc ( const std::string nameArg, const uint32_t fs, const uint32_t bufferSize_s, bool in_doNormalize = true, uint64_t in_normalizeValue = MAXCODABLEVALUE ) : TDSProcessor<double> (nameArg, fs, fs, bufferSize_s, _inputProc) {
 				this->in_doNormalize = in_doNormalize;
 				this->in_normalizeValue = in_normalizeValue;
 			}
@@ -45,14 +45,14 @@ namespace openAFE {
 			 * done here and the results are stocked in that private memory zone.
 			 * However, the results are not publiched yet on the output vectors.
 			 */
-			void processChunk (float* inChunkLeft, size_t leftDim, float* inChunkRight, size_t rightDim ) {
+			void processChunk (double* inChunkLeft, size_t leftDim, double* inChunkRight, size_t rightDim ) {
 				
 				// Appending the chunk to process (the processing must be done on the PMZ)
 				leftPMZ->appendTChunk( inChunkLeft, leftDim );
 				rightPMZ->appendTChunk( inChunkRight, rightDim );
 				
-				std::shared_ptr<twoCTypeBlock<float> > l_PMZ = leftPMZ->getLastChunkAccesor();
-				std::shared_ptr<twoCTypeBlock<float> > r_PMZ = rightPMZ->getLastChunkAccesor();
+				std::shared_ptr<twoCTypeBlock<double> > l_PMZ = leftPMZ->getLastChunkAccesor();
+				std::shared_ptr<twoCTypeBlock<double> > r_PMZ = rightPMZ->getLastChunkAccesor();
 
 				if ( this-> in_doNormalize ) {
 					// 0- Initialization
@@ -61,10 +61,10 @@ namespace openAFE {
 					unsigned long dim1_r = r_PMZ->array1.second;
 					unsigned long dim2_r = r_PMZ->array2.second;
 								
-					float* firstValue1_l = l_PMZ->array1.first;
-					float* firstValue2_l = l_PMZ->array2.first;
-					float* firstValue1_r = r_PMZ->array1.first;
-					float* firstValue2_r = r_PMZ->array2.first;				
+					double* firstValue1_l = l_PMZ->array1.first;
+					double* firstValue2_l = l_PMZ->array2.first;
+					double* firstValue1_r = r_PMZ->array1.first;
+					double* firstValue2_r = r_PMZ->array2.first;				
 						
 					if ( ( dim1_l > 0 ) && ( dim1_r > 0 ) ) {
 						std::thread leftThread( &InputProc::process, this, firstValue1_l, dim1_l );
